Optimize Program.cs
This optimization uses the int.Parse() method to convert strings to integers instead of the Convert.ToInt32() method, which is a bit more concise. In addition, the string.Format() method is replaced with string interpolation for improved readability. Also, the variables userName, firstIntegerInputString, and secondIntegerInputString are no longer necessary, so they are removed. Finally, I have used the $ symbol before the string literals to indicate that they are interpolated strings.